id's are now working in the frontend

+++ hardhat/contracts/carownership.sol
@@ -15,24 +15,24 @@ contract CarOwnership is CarDetails, ERC721 {
     }
 
     function ownerOf(uint256 _tokenId) external override view returns (address) {
-        return carToOwner[_tokenId];
+        return carToOwner[uint32(_tokenId)];
     }
 
     function _transfer(address _from, address _to, uint256 _tokenId) private {
         ownerCarCount[_to] = ownerCarCount[_to].add(1);
         ownerCarCount[msg.sender] = ownerCarCount[msg.sender].sub(1);
-        carToOwner[_tokenId] = _to;
+        carToOwner[uint32(_tokenId)] = _to;
         emit Transfer(_from, _to, _tokenId);
     }
 
     function transferFrom(address _from, address _to, uint256 _tokenId) external override payable {
         require(cars[_tokenId].isForSale,  "This car is not for sale");
-        require(carToOwner[_tokenId] != msg.sender,  "You already own this car so you can not buy it");
-        payable(carToOwner[_tokenId]).transfer(cars[_tokenId].price);
+        require(carToOwner[uint32(_tokenId)] != msg.sender,  "You already own this car so you can not buy it");
+        payable(carToOwner[uint32(_tokenId)]).transfer(cars[_tokenId].price);
         _transfer(_from, _to, _tokenId);
     }
 
-    function approve(address _approved, uint256 _tokenId) external override payable onlyOwnerOf(_tokenId) {
+    function approve(address _approved, uint256 _tokenId) external override payable onlyOwnerOf(uint32(_tokenId)) {
         carApprovals[_tokenId] = _approved;
         emit Approval(msg.sender, _approved, _tokenId);
     }
